apps/api/pyproject.toml
[project]
name = "atspro-api"
version = "0.1.0"
description = "Backend API for the ATSPro app"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "google-genai>=1.2.0",
    "httpx>=0.28.1",
    "openai>=1.65.4",
    "openai-agents>=0.0.14",
    "python-multipart>=0.0.20",
    "unstructured[doc,docx,md,pdf,rtf]>=0.16.23",
    "vlmrun-hub>=0.1.30",
    "psycopg[binary,pool]>=3.1.0",
    "redis>=5.0.0",
    "python-arango>=7.5.0",
    "rq>=1.15.0",
    "sse-starlette>=1.6.0",
    "alembic>=1.12.0",
    "pydantic-settings>=2.0.0",
    "psycopg-pool>=3.2.6",
]

[dependency-groups]
dev = [
    "ruff>=0.9.9",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-asyncio>=1.1.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["node_modules*", "tests*", "*.egg-info*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E203", "E501"]

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
